+++ src/main/java/com/lothrazar/cyclic/block/generatorfluid/ScreenGeneratorFluid.java
@@ -59,9 +59,9 @@ public class ScreenGeneratorFluid extends ScreenBase<ContainerGeneratorFluid> {
     energy.renderHoveredToolTip(ms, mouseX, mouseY, container.tile.getEnergy());
     progress.renderHoveredToolTip(ms, mouseX, mouseY, container.tile.getField(TileGeneratorFluid.Fields.TIMER.ordinal()));
     ms.push();
-    ms.translate(this.width / 2, this.height / 2, 0);
+    ms.translate((double) this.width / 2, (double) this.height / 2, 0);
     ms.rotate(Vector3f.ZP.rotationDegrees(-90));
-    ms.translate(-this.width / 2, -this.height / 2, 0);
+    ms.translate((double) -this.width / 2, (double) -this.height / 2, 0);
     Vector4f vec = new Vector4f(mouseX, mouseY, 0, 1);
     vec.transform(ms.getLast().getMatrix());
     ms.pop(); //Look, it's a bit hacky, but it gets the job done.  Rotation Math!
@@ -87,9 +87,9 @@ public class ScreenGeneratorFluid extends ScreenBase<ContainerGeneratorFluid> {
     progress.max = container.tile.getField(TileGeneratorFluid.Fields.BURNMAX.ordinal());
     progress.draw(ms, container.tile.getField(TileGeneratorFluid.Fields.TIMER.ordinal()));
     ms.push();
-    ms.translate(this.width / 2, this.height / 2, 0);
+    ms.translate((double) this.width / 2, (double) this.height / 2, 0);
     ms.rotate(Vector3f.ZP.rotationDegrees(90));
-    ms.translate(-this.width / 2, -this.height / 2, 0);
+    ms.translate((double) -this.width / 2, (double) -this.height / 2, 0);
     fluid.draw(ms, container.tile.getFluid());
     ms.pop();
   }
